requirements.txt
torch
gsplat==1.5.3
numpy
Pillow
packaging
ninja
PyYAML
imageio
tqdm
torchmetrics
PyOpenGL
pyopengltk
addict
bridge-python-sdk @ git+https://github.com/Looking-Glass/bridge-python-sdk
huggingface-hub==0.34.0
transformers==4.40.2
-e tools/sharp
